requirements.txt
google-play-scraper
pandas
nltk
scikit-learn
psycopg2-binary

